// resources/js/types/models/Sprint.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Sprint = {
    id: number;
    title: string | null;
    start_time: string | null;
    end_time: string | null;
    duration: number;
    is_active: boolean;
    board_id: number;
    created_at: string | null;
    updated_at: string | null;
};